+++ pkg/cpu/doc.go
@@ -1,0 +1,2 @@
+// Package cpu provides utilities for CPU usage.
+package cpu
