+++ pkg/cpu/utils_test.go
