+++ pkg/memory/utils.go
