+++ pkg/memory/utils_test.go
